# backend_contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/files/{id}": {
      "delete": {
        "operationId": "FilesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "FilesController_patch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/profile": {
      "get": {
        "operationId": "ProfileController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "ProfileController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/subjects/{id}/files": {
      "post": {
        "operationId": "SubjectsController_addFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Aqademiq API",
    "description": "NestJS backend — /v1, snake_case. Auth: custom RS256 JWT (Bearer).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PatchFileDto": {
        "type": "object",
        "properties": {
          "important": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          }
        }
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 0,
            "maxLength": 120
          },
          "gender": {
            "type": "string",
            "minLength": 0,
            "maxLength": 40
          },
          "date_of_birth": {
            "type": "string",
            "pattern": "/^\\d{4}-\\d{2}-\\d{2}$/"
          },
          "university": {
            "type": "string",
            "minLength": 0,
            "maxLength": 200
          },
          "program": {
            "type": "string",
            "minLength": 0,
            "maxLength": 200
          },
          "avatar_index": {
            "type": "number",
            "minimum": 0,
            "maximum": 7
          }
        }
      },
...
    }
  }
}